refactor(*): properly split unix support into its own folder
