+++ src/kerberos_common.h
@@ -2,7 +2,7 @@
 #define COMMON_H
 
 #include <nan.h>
-#include "kerberos_gss.h"
+#include "unix/kerberos_gss.h"
 
 // Provide a default custom delter for the `gss_result` type
 inline void ResultDeleter(gss_result* result) {
