+++ src/kerberos_server.cc
@@ -1,7 +1,4 @@
-#include <memory>
-
 #include "kerberos_server.h"
-#include "kerberos_worker.h"
 
 Nan::Persistent<v8::Function> KerberosServer::constructor;
 NAN_MODULE_INIT(KerberosServer::Init) {
@@ -36,13 +33,6 @@ v8::Local<v8::Object> KerberosServer::NewInstance(gss_server_state* state) {
 
 KerberosServer::KerberosServer(gss_server_state* state) : _state(state) {}
 
-KerberosServer::~KerberosServer() {
-    if (_state != NULL) {
-        authenticate_gss_server_clean(_state);
-        _state = NULL;
-    }
-}
-
 gss_server_state* KerberosServer::state() const {
     return _state;
 }
@@ -71,32 +61,4 @@ NAN_GETTER(KerberosServer::TargetNameGetter) {
 NAN_GETTER(KerberosServer::ContextCompleteGetter) {
     KerberosServer* server = Nan::ObjectWrap::Unwrap<KerberosServer>(info.This());
     info.GetReturnValue().Set(Nan::New(server->_state->context_complete));
-}
-
-NAN_METHOD(KerberosServer::Step) {
-    KerberosServer* server = Nan::ObjectWrap::Unwrap<KerberosServer>(info.This());
-    std::string challenge(*Nan::Utf8String(info[0]));
-    Nan::Callback* callback = new Nan::Callback(Nan::To<v8::Function>(info[1]).ToLocalChecked());
-
-    KerberosWorker::Run(callback, "kerberos:ServerStep", [=](KerberosWorker::SetOnFinishedHandler onFinished) {
-        std::shared_ptr<gss_result> result(
-            authenticate_gss_server_step(server->state(), challenge.c_str()), ResultDeleter);
-
-        return onFinished([=](KerberosWorker* worker) {
-            Nan::HandleScope scope;
-            if (result->code == AUTH_GSS_ERROR) {
-                v8::Local<v8::Value> argv[] = {Nan::New(result->message).ToLocalChecked(), Nan::Null()};
-                worker->Call(2, argv);
-                return;
-            }
-
-            v8::Local<v8::Value> response = Nan::Null();
-            if (server->state()->response != NULL) {
-                response = Nan::New(server->state()->response).ToLocalChecked();
-            }
-
-            v8::Local<v8::Value> argv[] = {Nan::Null(), response};
-            worker->Call(2, argv);
-        });
-    });
 }
